+++ Block/Stub.php
@@ -30,6 +30,9 @@ namespace Nosto\Tagging\Block;
 use Magento\Framework\View\Element\Template;
 use Nosto\Tagging\Helper\Account as NostoHelperAccount;
 use Nosto\Tagging\Helper\Scope as NostoHelperScope;
+use Nosto\Tagging\Helper\Data as NostoHelperData;
+use Nosto\Tagging\Helper\Variation as NostoHelperVariation;
+use Nosto\Tagging\Helper\Customer as NostoHelperCustomer;
 
 /**
  * Nosto JS stub block
@@ -45,20 +48,44 @@ class Stub extends Template
     }
 
     /**
+     * @var NostoHelperData
+     */
+    private $nostoHelperData;
+
+    /**
+     * @var NostoHelperCustomer
+     */
+    private $nostoHelperCustomer;
+
+    /**
+     * @var NostoHelperVariation
+     */
+    private $nostoHelperVariation;
+
+    /**
      * Stub constructor.
-     * @param Template\Context $context the context.
+     * @param Template\Context $context
      * @param NostoHelperAccount $nostoHelperAccount
      * @param NostoHelperScope $nostoHelperScope
-     * @param array $data optional data.
+     * @param NostoHelperData $nostoHelperData
+     * @param NostoHelperCustomer $nostoHelperCustomer
+     * @param NostoHelperVariation $nostoHelperVariation
+     * @param array $data
      */
     public function __construct(
         Template\Context $context,
         NostoHelperAccount $nostoHelperAccount,
         NostoHelperScope $nostoHelperScope,
+        NostoHelperData $nostoHelperData,
+        NostoHelperCustomer $nostoHelperCustomer,
+        NostoHelperVariation $nostoHelperVariation,
         array $data = []
     ) {
         parent::__construct($context, $data);
         $this->taggingConstruct($nostoHelperAccount, $nostoHelperScope);
+        $this->nostoHelperData = $nostoHelperData;
+        $this->nostoHelperCustomer = $nostoHelperCustomer;
+        $this->nostoHelperVariation = $nostoHelperVariation;
     }
 
     /**
@@ -68,5 +95,26 @@ class Stub extends Template
     public function getAbstractObject()
     {
         return null;
+    }
+
+    /**
+     * Returns if autoloading recommendations is disabled or not.
+     *
+     * @return boolean
+     */
+    public function isRecoAutoloadDisabled()
+    {
+        $store = $this->getNostoHelperScope()->getStore(true);
+        // If price variations are used and the variation something else than
+        // the default one we disable the autoload. For default variation
+        // the sections are not loaded and loadRecommendations() is not called
+        if ($this->nostoHelperData->isPricingVariationEnabled($store)
+            && !$this->nostoHelperVariation->isDefaultVariationCode(
+                $this->nostoHelperCustomer->getGroupCode()
+            )
+        ) {
+            return true;
+        }
+        return false;
     }
 }