+++ Block/Variation.php
@@ -32,9 +32,10 @@ use Magento\Framework\View\Element\Template\Context;
 use Nosto\Tagging\Helper\Account as NostoHelperAccount;
 use Nosto\Tagging\Helper\Currency as NostoHelperCurrency;
 use Nosto\Tagging\Helper\Scope as NostoHelperScope;
-use Nosto\Object\MarkupableString;
 use Nosto\Tagging\Helper\Data as NostoHelperData;
 use Nosto\Tagging\Helper\Customer as NostoHelperCustomer;
+use Nosto\Tagging\Helper\Variation as NostoHelperVariation;
+use Nosto\Object\MarkupableString;
 
 /**
  * Page type block used for outputting the variation identifier on the different pages.
@@ -45,18 +46,36 @@ class Variation extends Template
         TaggingTrait::__construct as taggingConstruct; // @codingStandardsIgnoreLine
     }
 
+    /**
+     * @var NostoHelperCurrency
+     */
     private $nostoHelperCurrency;
+
+    /**
+     * @var NostoHelperData
+     */
     private $nostoHelperData;
+
+    /**
+     * @var NostoHelperCustomer
+     */
     private $nostoHelperCustomer;
 
     /**
-     * Variation block constructor.
+     * @var NostoHelperVariation
+     */
+    private $nostoHelperVariation;
+
+    /**
+     * Variation constructor.
      *
      * @param Context $context
      * @param NostoHelperAccount $nostoHelperAccount
      * @param NostoHelperScope $nostoHelperScope
      * @param NostoHelperCurrency $nostoHelperCurrency
      * @param NostoHelperData $nostoHelperData
+     * @param NostoHelperCustomer $nostoHelperCustomer
+     * @param NostoHelperVariation $nostoHelperVariation
      * @param array $data
      */
     public function __construct(
@@ -66,6 +85,7 @@ class Variation extends Template
         NostoHelperCurrency $nostoHelperCurrency,
         NostoHelperData $nostoHelperData,
         NostoHelperCustomer $nostoHelperCustomer,
+        NostoHelperVariation $nostoHelperVariation,
         array $data = []
     ) {
         parent::__construct($context, $data);
@@ -74,6 +94,7 @@ class Variation extends Template
         $this->nostoHelperCurrency = $nostoHelperCurrency;
         $this->nostoHelperData = $nostoHelperData;
         $this->nostoHelperCustomer = $nostoHelperCustomer;
+        $this->nostoHelperVariation = $nostoHelperVariation;
     }
 
     /**
@@ -114,8 +135,16 @@ class Variation extends Template
     {
         $store = $this->nostoHelperScope->getStore(true);
 
+        // We inject the active variation tag if the exchange rates are used or
+        // if the price variations are used and the active variation is the
+        // default one
         if ($this->nostoHelperCurrency->exchangeRatesInUse($store)
-            || $this->nostoHelperData->isPricingVariationEnabled($store)
+            || ($this->nostoHelperData->isPricingVariationEnabled($store)
+                && $this->nostoHelperVariation->isDefaultVariationCode(
+                    $this->nostoHelperCustomer->getGroupCode()
+                )
+
+            )
         ) {
             return new MarkupableString(
                 $this->getVariationId(),